+++ bott/loss.py
@@ -3,25 +3,24 @@ import torch
 
 class LossFunction(torch.nn.Module):
 
-    def __init__(self, loss_params, y_true, device='cuda'):
+    def __init__(self, loss_params, device='cuda'):
         super(LossFunction, self).__init__()
         self.device = device
         self.loss_params = loss_params
-        self.y_true = y_true
         
-    def forward(self, y_simu):
+    def forward(self, y_simu, y_true, reduce=True):
         loss_params = self.loss_params
         loss_type   = loss_params['loss_type']
         dp_pow      = loss_params.get('dp_pow', 1)
         
         if loss_type == 'SSE':
-            loss = SSE(y_simu, self.y_true, dp_pow)
+            loss = SSE(y_simu, y_true, dp_pow, reduce=reduce)
         elif loss_type == 'MSE':
-            loss = MSE(y_simu, self.y_true, dp_pow)
+            loss = MSE(y_simu, y_true, dp_pow, reduce=reduce)
         elif loss_type == 'NRMSE':
-            loss = NRMSE(y_simu, self.y_true, dp_pow)
+            loss = NRMSE(y_simu, y_true, dp_pow, reduce=reduce)
         else:
-            raise ValueError(f"The current implementation does not support tiling type {loss_type}")
+            raise ValueError(f"The current implementation does not support lossing type {loss_type}")
         
         return loss
     
@@ -41,21 +40,34 @@ I'm still unsure what would be the appropriate shape for our loss yet
 '''
 
 # Loss functions
-def SSE(y_simu, y_true, dp_pow):
+def SSE(y_simu, y_true, dp_pow, reduce=True):
+    # reduce decides whether we reduce the batch dimension
     y_simu = safe_power(y_simu, dp_pow)
     y_true = safe_power(y_true, dp_pow)
-    return (y_simu-y_true).pow(2).sum(dim=-1).sum(dim=-1)
+    if reduce:
+        reduce_dims = tuple(range(y_simu.ndim))
+    else: 
+        reduce_dims = tuple(range(1, y_simu.ndim))
+    return (y_simu-y_true).pow(2).sum(dim=reduce_dims)
 
-def MSE(y_simu, y_true, dp_pow):
+def MSE(y_simu, y_true, dp_pow, reduce=True):
     y_simu = safe_power(y_simu, dp_pow)
     y_true = safe_power(y_true, dp_pow)
-    return (y_simu-y_true).pow(2).mean()
+    if reduce:
+        reduce_dims = tuple(range(y_simu.ndim))
+    else: 
+        reduce_dims = tuple(range(1, y_simu.ndim))
+    return (y_simu-y_true).pow(2).mean(dim=reduce_dims)
 
-def NRMSE(y_simu, y_true, dp_pow):
+def NRMSE(y_simu, y_true, dp_pow, reduce=True):
     y_simu = safe_power(y_simu, dp_pow)
     y_true = safe_power(y_true, dp_pow)
     data_mean = y_true.mean()
-    return (y_simu-y_true).pow(2).mean().sqrt() / data_mean
+    if reduce:
+        reduce_dims = tuple(range(y_simu.ndim))
+    else: 
+        reduce_dims = tuple(range(1, y_simu.ndim))
+    return (y_simu-y_true).pow(2).mean(dim=reduce_dims).sqrt() / data_mean
 
 def safe_power(arr, power, eps=1e-6):
     '''